[BUGFIX] Simplify SimpleDOM casting
The new `cast` function added for the recent TS upgrade has a couple of
issues:

1. It creates wrapping objects which modify the values that are
   wrapped/casted. Removing `cast()`, as the production build currently
   does, breaks because of this.
2. These wrapping objects conflate casting _to_ SimpleDOM and casting
   _away_ from it, into browser APIs.

The new system makes `cast` into two pure validation utility functions,
which either downlevels to SimpleDOM (without any checks, as SimpleDOM
is a subset of the real DOM), or uplevels to DOM with appropriate checks
(e.g. are we actually in the browser? Is the node a DOM node? Is it the
expected type of DOM node?).